VMware: Split out VMwareAPISession
The object is not only used by the driver,
but in practically all modules of vmwareapi.

It reduces a bit the scope of the driver module
itself
