+++ nova/tests/unit/virt/vmwareapi/test_vmops.py
@@ -44,9 +44,9 @@ from nova import version
 from nova.virt import hardware
 from nova.virt.vmwareapi import cluster_util
 from nova.virt.vmwareapi import constants
-from nova.virt.vmwareapi import driver
 from nova.virt.vmwareapi import ds_util
 from nova.virt.vmwareapi import images
+from nova.virt.vmwareapi.session import VMwareAPISession
 from nova.virt.vmwareapi import vif
 from nova.virt.vmwareapi import vim_util
 from nova.virt.vmwareapi import vm_util
@@ -75,7 +75,7 @@ class VMwareVMOpsTestCase(test.TestCase):
                    my_ip='',
                    flat_injected=True)
         self._context = context.RequestContext('fake_user', 'fake_project')
-        self._session = driver.VMwareAPISession()
+        self._session = VMwareAPISession()
 
         self._virtapi = mock.Mock()
         self._image_id = nova.tests.unit.image.fake.get_valid_image_id()
